FISH-315 Update Eventbus Notifier to Use New API
