+++ nucleus/payara-modules/notification-eventbus-core/src/main/java/fish/payara/notification/eventbus/core/EventbusNotifierConfiguration.java
@@ -1,6 +1,6 @@
 /*
  *
- * Copyright (c) 2016-2017 Payara Foundation and/or its affiliates. All rights reserved.
+ * Copyright (c) [2016-2020] Payara Foundation and/or its affiliates. All rights reserved.
  *
  * The contents of this file are subject to the terms of either the GNU
  * General Public License Version 2 only ("GPL") or the Common Development
@@ -38,11 +38,10 @@
  */
 package fish.payara.notification.eventbus.core;
 
-import fish.payara.nucleus.notification.configuration.NotifierConfiguration;
-import fish.payara.nucleus.notification.configuration.NotifierConfigurationType;
-import fish.payara.nucleus.notification.configuration.NotifierType;
 import org.jvnet.hk2.config.Attribute;
 import org.jvnet.hk2.config.Configured;
+
+import fish.payara.internal.notification.PayaraNotifierConfiguration;
 
 import java.beans.PropertyVetoException;
 
@@ -50,8 +49,7 @@ import java.beans.PropertyVetoException;
  * @author mertcaliskan
  */
 @Configured
-@NotifierConfigurationType(type = NotifierType.EVENTBUS)
-public interface EventbusNotifierConfiguration extends NotifierConfiguration {
+public interface EventbusNotifierConfiguration extends PayaraNotifierConfiguration {
 
     @Attribute(defaultValue = "payara.notification.event")
     String getTopicName();
