+++ nucleus/payara-modules/notification-eventbus-core/src/main/java/fish/payara/notification/eventbus/core/GetEventbusNotifierConfiguration.java
@@ -1,7 +1,7 @@
 /*
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS HEADER.
  *
- * Copyright (c) 2017 Payara Foundation and/or its affiliates. All rights reserved.
+ * Copyright (c) [2017-2020] Payara Foundation and/or its affiliates. All rights reserved.
  *
  * The contents of this file are subject to the terms of either the GNU
  * General Public License Version 2 only ("GPL") or the Common Development
@@ -39,16 +39,20 @@
  */
 package fish.payara.notification.eventbus.core;
 
-import com.sun.enterprise.util.ColumnFormatter;
-import fish.payara.nucleus.notification.admin.BaseGetNotifierConfiguration;
-import fish.payara.nucleus.notification.configuration.NotificationServiceConfiguration;
-import java.util.HashMap;
 import java.util.Map;
-import org.glassfish.api.admin.*;
+
+import org.glassfish.api.admin.CommandLock;
+import org.glassfish.api.admin.ExecuteOn;
+import org.glassfish.api.admin.RestEndpoint;
+import org.glassfish.api.admin.RestEndpoints;
+import org.glassfish.api.admin.RuntimeType;
 import org.glassfish.config.support.CommandTarget;
 import org.glassfish.config.support.TargetType;
 import org.glassfish.hk2.api.PerLookup;
 import org.jvnet.hk2.annotations.Service;
+
+import fish.payara.internal.notification.admin.BaseGetNotifierConfiguration;
+import fish.payara.internal.notification.admin.NotificationServiceConfiguration;
 
 /**
  * @author mertcaliskan
@@ -67,29 +71,22 @@ import org.jvnet.hk2.annotations.Service;
 public class GetEventbusNotifierConfiguration extends BaseGetNotifierConfiguration<EventbusNotifierConfiguration> {
 
     @Override
-    protected String listConfiguration(EventbusNotifierConfiguration configuration) {
-        String headers[] = {"Enabled", "Noisy", "Topic Name"};
-        ColumnFormatter columnFormatter = new ColumnFormatter(headers);
-        Object values[] = new Object[3];
+    protected Map<String, Object> getNotifierConfiguration(EventbusNotifierConfiguration configuration) {
+        Map<String, Object> map = super.getNotifierConfiguration(configuration);
 
-        values[0] = configuration.getEnabled();
-        values[1] = configuration.getNoisy();
-        values[2] = configuration.getTopicName();
+        if (configuration != null) {
+            map.put("Topic Name", configuration.getTopicName());
+        }
 
-        columnFormatter.addRow(values);
-        return columnFormatter.toString();
+        return map;
     }
 
     @Override
-    protected Map<String, Object> getNotifierConfiguration(EventbusNotifierConfiguration configuration) {
-        Map<String, Object> map = new HashMap<>(3);
+    protected Map<String, Object> getNotifierProperties(EventbusNotifierConfiguration configuration) {
+        Map<String, Object> map = super.getNotifierConfiguration(configuration);
 
         if (configuration != null) {
-            map.put("enabled", configuration.getEnabled());
-            map.put("noisy", configuration.getNoisy());
             map.put("topicName", configuration.getTopicName());
-        } else {
-            map.put("noisy", Boolean.TRUE.toString());
         }
 
         return map;
